Make time delta 16 bits wide
The delta is in milliseconds, which corresponds to a max time delta
of 65.5 s.  Actual deltas will be far below this limit, so it's safe
to use a uint16_t for delta.

+++ src/alien.cpp
@@ -34,7 +34,7 @@
 
 namespace GameEntities {
 
-    void Alien::movement(Uint32 delta)
+    void Alien::movement(uint16_t delta)
     {
         // control in place animation
         frame_time_count += delta;
